Systemdbrestore

+++ internal/Restore-DBFromFilteredArray.ps1
@@ -36,8 +36,9 @@ Function Restore-DBFromFilteredArray {
         [switch]$Continue,
         [string]$AzureCredential,
         [switch]$ReplaceDbNameInFile,
-        [string]$OldDatabaseName,
-        [string]$DestinationFileSuffix
+        [string]$DestinationFileSuffix,
+		[string[]]$DatabaseFilter,
+		[Bool]$SystemRestore
     )
     begin {
         $FunctionName = (Get-PSCallstack)[0].Command
@@ -83,22 +84,19 @@ Function Restore-DBFromFilteredArray {
             $DestinationLogDirectory = Get-SqlDefaultPaths $Server log
         }
 
-        If ($DbName -in $Server.databases.name) {
-            if (($ScriptOnly -eq $true) -or ($verifyonly -eq $true)) {
-                Write-Message -Level Verbose -Message "No need to close db for this operation"
-            }
-            elseIf ($WithReplace -eq $true -and $VerifyOnly -eq $false) {
-                if ($Pscmdlet.ShouldProcess("Killing processes in $dbname on $SqlInstance as it exists and WithReplace specified  `n", "Cannot proceed if processes exist, ", "Database Exists and WithReplace specified, need to kill processes to restore")) {
-                    try {
-                        Write-Message -Level Verbose -Message "Set $DbName single_user to kill processes"
-                        Stop-DbaProcess -SqlInstance $Server -Databases $Dbname -WarningAction Silentlycontinue
-                        if ($Continue -eq $false) {
-                            $server.Query("Alter database $DbName set offline with rollback immediate; alter database $DbName set restricted_user; Alter database $DbName set online with rollback immediate",'master')
+        If ($DbName -in $Server.databases.name -and ($ScriptOnly -eq $false -or $VerfiyOnly -eq $false)) {
+            If ($ReplaceDatabase -eq $true) {
+                if (!($SystemRestore)) {
+                    if ($Pscmdlet.ShouldProcess("Killing processes in $dbname on $SqlInstance as it exists and WithReplace specified  `n", "Cannot proceed if processes exist, ", "Database Exists and WithReplace specified, need to kill processes to restore")) {
+                        try {
+                            Write-Message -Level Verbose -Message "Set $DbName single_user to kill processes"
+                            Stop-DbaProcess -SqlInstance $Server -Database $Dbname -WarningAction Silentlycontinue
+                            Invoke-DbaSqlcmd -ServerInstance:$SqlInstance -Credential:$SqlCredential -query "Alter database $DbName set offline with rollback immediate; alter database $DbName set restricted_user; Alter database $DbName set online with rollback immediate" -database master
+                            $server.ConnectionContext.Connect()
                         }
-                        $server.ConnectionContext.Connect()
-                    }
-                    catch {
-                        Write-Message -Level Verbose -Message "No processes to kill in $DbName"
+                        catch {
+                            Write-Message -Level Verbose -Message "No processes to kill in $DbName"
+                        }
                     }
                 }
             }
@@ -279,8 +277,7 @@ Function Restore-DBFromFilteredArray {
             elseif ($RestoreFiles[0].RecoveryModel -ne 'Simple') {
                 $Restore.ToPointInTime = $RestoreTime
                 Write-Message -Level Verbose -Message "restoring to $RestoreTime"
-
-            }
+            } 
             else {
                 Write-Message -Level Verbose -Message "Restoring a Simple mode db, no restoretime"
             }
@@ -362,8 +359,9 @@ Function Restore-DBFromFilteredArray {
                     Write-Message -Level Verbose -Message "Failed, Closing Server connection"
                     $RestoreComplete = $False
                     $ExitError = $_.Exception.InnerException
-					Stop-Function -Message "Failed to restore db $DbName, stopping" -ErrorRecord $_
-					return
+                    Write-Warning "$FunctionName - $ExitError" -WarningAction stop
+                    #Exit as once one restore has failed there's no point continuing
+                    break
                 }
                 finally {
                     if ($ReuseSourceFolderStructure) {
